feat: use grcp load balancing; feat: use async autoscaler for instances;
This includes a big refactor of the Call method inside the worker.
The worker now maintains a load balancer for each registered function.
I benchmarked this using the tests/leaf/main.go program and it looks like a VERY promising approach in comparison to the old one. ~30ms call latency @ 600 RPS .

TODOs:
- improve the proto definitions for the Call endpoint (we now only need functionID)
- delete the old state management in the leaf in favor of the new small state (rename this to just state)
- improve naming of components in leaf
- decide what to do with docker network. Do we really need it?